+++ src/love/script.cpp
@@ -13,10 +13,9 @@ using love::Types::Graphics::Image;
 using love::Types::Graphics::Font;
 using love::Types::Graphics::Point;
 using love::Types::Graphics::Color;
-using love::Types::System::Config;
 using love::Types::Input::Joystick;
-using love::Types::System::windowConfig;
-using love::Types::System::moduleConfig;
+using love::Types::Config::WindowConfig;
+using love::Types::Config::ModuleConfig;
 using love::Types::Audio::SoundData;
 using love::graphics;
 
@@ -71,8 +70,10 @@ script::script(const std::string& file) {
 	chai.register_namespace([](chaiscript::Namespace& love) {
 		ChaiLove* app = ChaiLove::getInstance();
 		love["audio"] = var(std::ref(app->audio));
+		love["config"] = var(std::ref(app->config));
 		love["event"] = var(std::ref(app->event));
 		love["filesystem"] = var(std::ref(app->filesystem));
+		love["font"] = var(std::ref(app->font));
 		love["graphics"] = var(std::ref(app->graphics));
 		love["image"] = var(std::ref(app->image));
 		love["joystick"] = var(std::ref(app->joystick));
@@ -139,23 +140,22 @@ script::script(const std::string& file) {
 	chai.add(fun<int, Font, const std::string&>(&Font::getWidth), "getWidth");
 
 	// Config
-	chai.add(user_type<windowConfig>(), "windowConfig");
-	chai.add(fun(&windowConfig::width), "width");
-	chai.add(fun(&windowConfig::height), "height");
-	chai.add(fun(&windowConfig::bbp), "bbp");
-	chai.add(fun(&windowConfig::title), "title");
-	chai.add(fun(&windowConfig::asyncblit), "asyncblit");
-	chai.add(fun(&windowConfig::hwsurface), "hwsurface");
-	chai.add(fun(&windowConfig::doublebuffering), "doublebuffering");
-	chai.add(user_type<moduleConfig>(), "moduleConfig");
-	chai.add(fun(&moduleConfig::sound), "sound");
-	chai.add(user_type<Config>(), "Config");
-	chai.add(fun(&Config::identity), "identity");
-	chai.add(fun(&Config::version), "version");
-	chai.add(fun(&Config::window), "window");
-	chai.add(fun(&Config::modules), "modules");
-	chai.add(fun(&Config::options), "options");
-	chai.add_global(var(std::ref(app->config)), "config");
+	chai.add(user_type<WindowConfig>(), "WindowConfig");
+	chai.add(fun(&WindowConfig::width), "width");
+	chai.add(fun(&WindowConfig::height), "height");
+	chai.add(fun(&WindowConfig::bbp), "bbp");
+	chai.add(fun(&WindowConfig::title), "title");
+	chai.add(fun(&WindowConfig::asyncblit), "asyncblit");
+	chai.add(fun(&WindowConfig::hwsurface), "hwsurface");
+	chai.add(fun(&WindowConfig::doublebuffering), "doublebuffering");
+	chai.add(user_type<ModuleConfig>(), "ModuleConfig");
+	chai.add(fun(&ModuleConfig::sound), "sound");
+	chai.add(user_type<config>(), "Config");
+	chai.add(fun(&config::identity), "identity");
+	chai.add(fun(&config::version), "version");
+	chai.add(fun(&config::window), "window");
+	chai.add(fun(&config::modules), "modules");
+	chai.add(fun(&config::options), "options");
 
 	// Joystick
 	chai.add(user_type<Joystick>(), "Joystick");
@@ -206,26 +206,21 @@ script::script(const std::string& file) {
 	chai.add(fun<love::graphics&, graphics, int, int, int, int>(&graphics::clear), "clear");
 	chai.add(fun<love::graphics&, graphics, int, int, int>(&graphics::clear), "clear");
 	chai.add(fun<love::graphics&, graphics>(&graphics::clear), "clear");
-	chai.add_global(var(std::ref(app->graphics)), "graphics");
 
 	// Font
 	chai.add(fun(&font::isOpen), "isOpen");
-	chai.add_global(var(std::ref(app->font)), "font");
 
 	// Keyboard
 	chai.add(fun<bool, keyboard, const std::string&>(&keyboard::isDown), "isDown");
 	chai.add(fun(&keyboard::setKeyRepeat), "setKeyRepeat");
 	chai.add(fun(&keyboard::getKeyFromScancode), "getKeyFromScancode");
 	chai.add(fun(&keyboard::getScancodeFromKey), "getScancodeFromKey");
-	chai.add_global(var(std::ref(app->keyboard)), "keyboard");
 
 	// Event
 	chai.add(fun(&event::quit), "quit");
-	chai.add_global(var(std::ref(app->event)), "event");
 
 	// Image
 	chai.add(fun(&image::newImageData), "newImageData");
-	chai.add_global(var(std::ref(app->image)), "image");
 
 	// Filesystem
 	chai.add(fun(&filesystem::unmount), "unmount");
@@ -239,13 +234,11 @@ script::script(const std::string& file) {
 	chai.add(fun<std::vector<std::string>, filesystem, const std::string&>(&filesystem::lines), "lines");
 	chai.add(fun<std::vector<std::string>, filesystem, const std::string&, const std::string&>(&filesystem::lines), "lines");
 	chai.add(fun(&filesystem::load), "load");
-	chai.add_global(var(std::ref(app->filesystem)), "filesystem");
 
 	// System
 	chai.add(fun(&system::getOS), "getOS");
 	chai.add(fun(&system::getVersion), "getVersion");
 	chai.add(fun(&system::getVersionString), "getVersionString");
-	chai.add_global(var(std::ref(app->system)), "system");
 
 	// Mouse
 	chai.add(fun(&mouse::setVisible), "setVisible");
@@ -257,29 +250,24 @@ script::script(const std::string& file) {
 	chai.add(fun(&mouse::getPosition), "getPosition");
 	chai.add(fun<bool, mouse, const std::string&>(&mouse::isDown), "isDown");
 	chai.add(fun<bool, mouse, int>(&mouse::isDown), "isDown");
-	chai.add_global(var(std::ref(app->mouse)), "mouse");
 
 	// Sound
 	chai.add(fun<SoundData*, sound, const std::string&>(&sound::newSoundData), "newSoundData");
-	chai.add_global(var(std::ref(app->sound)), "sound");
 
 	// Audio
 	chai.add(fun(&audio::play), "play");
 	chai.add(fun<SoundData*, audio, const std::string&>(&audio::newSource), "newSource");
 	chai.add(fun(&audio::getVolume), "getVolume");
 	chai.add(fun(&audio::setVolume), "setVolume");
-	chai.add_global(var(std::ref(app->audio)), "audio");
 
 	// Window
 	chai.add(fun(&window::setTitle), "setTitle");
 	chai.add(fun(&window::getTitle), "getTitle");
-	chai.add_global(var(std::ref(app->window)), "window");
 
 	// Timer
 	chai.add(fun(&timer::getDelta), "getDelta");
 	chai.add(fun(&timer::getFPS), "getFPS");
 	chai.add(fun(&timer::step), "step");
-	chai.add_global(var(std::ref(app->timer)), "timer");
 
 	// Joystick
 	chai.add(fun(&joystick::getJoysticks), "getJoysticks");
@@ -287,7 +275,6 @@ script::script(const std::string& file) {
 	chai.add(fun<bool, joystick, int, const std::string&>(&joystick::isDown), "isDown");
 	chai.add(fun<bool, joystick, int, int>(&joystick::isDown), "isDown");
 	chai.add(fun(&joystick::operator[]), "[]");
-	chai.add_global(var(std::ref(app->joystick)), "joystick");
 
 	// Math
 	auto mathlib = chaiscript::extras::math::bootstrap();
@@ -309,7 +296,6 @@ script::script(const std::string& file) {
 	chai.add(fun<std::string, math, const std::string&>(&math::compress), "compress");
 	chai.add(fun<std::string, math, const std::string&, int>(&math::compress), "compress");
 	chai.add(fun(&math::decompress), "decompress");
-	chai.add_global(var(std::ref(app->math)), "math");
 
 	// Ensure the love namespace is imported and ready.
 	chai.import("love");
@@ -348,7 +334,7 @@ script::script(const std::string& file) {
 		hasUpdate = false;
 	}
 	try {
-		chaiconf = chai.eval<std::function<void(Config&)> >("conf");
+		chaiconf = chai.eval<std::function<void(config&)> >("conf");
 	}
 	catch (const std::exception& e) {
 		std::cout << "[ChaiLove] [script] conf(t) " << e.what() << std::endl;
@@ -427,7 +413,7 @@ script::script(const std::string& file) {
 	#endif
 }
 
-void script::conf(Config& t) {
+void script::conf(config& t) {
 	#ifdef __HAVE_CHAISCRIPT__
 	if (hasconf) {
 		try {